Decouple CLI and service
